ml_pipeline/requirements.txt
azureml-sdk
azureml-defaults
azure-cli
tensorflow
sklearn
numpy
matplotlib
opencv-python
python-dotenv